requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0

streamlit==1.22.0
tabulate==0.9.0
timm==0.9.2
transformers==4.29.2

pillow
torch